more refactoring; fix bug that browser stays opened when script exits
